Add background subtraction
There are two options for background subtraction. A standard
subtraction routine (note that this will require modification of
thresholding values), or a matching subtraction.

The match algorithm works by setting matching pixels on the
background and current frame to white. The background image
is always the first frame of the video.